fix sparse neuromc add counting
